requirements.txt
beautifulsoup4==4.11.1
dateparser==1.1.1
feedparser==6.0.10
newspaper3k==0.2.8
nltk==3.7
requests==2.28.1
streamlit==1.12.2
